SnapshotStrategy should be able to access to the final state of the issue
When the field value was assigned at the moment of issue creation - there will be no activity for that assignment. To handle this, it's necessary to compare the final issue state with the state we recovered with activities. The fields that are different - was assignment at the moment of issue creation